requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Agent Framework
langgraph>=0.0.40
langchain-core>=0.1.0
langchain-groq>=0.1.0

# Backend
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0

# GitHub App authentication
PyJWT[crypto]>=2.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
